ecwid.ID type for object id`s
json can contains id for deleted item as -1.
unmarshaled as 0

+++ store_info_test.go
@@ -38,5 +38,5 @@ func (suite *StoreInformationTestSuite) TestStoreProfileGet() {
 	suite.Truef(requested, "request failed")
 
 	suite.Nil(err)
-	suite.Equal(uint64(storeID), p.GeneralInfo.StoreID, "id")
+	suite.Equal(storeID, p.GeneralInfo.StoreID, "id")
 }
